+++ model/alliancemodel/alliance_info_model.go
@@ -21,7 +21,7 @@ type AllianceInfoModel struct {
 	AllianceName       string  `json:"alliance_name"`
 	AllianceCountLimit int32   `json:"alliance_count_limit"` // 联盟中家族数量限制
 	FamilyIDs          []int32 `json:"family_ids"`           // 联盟中家族ID列表
-	AllianceGroupID    int32   `json:"alliance_group_id"`    // 联盟所属组ID
+	AllianceGroupID    int64   `json:"alliance_group_id"`    // 联盟所属组ID
 }
 
 func LoadAllianceInfoModel(ctx context.Context, allianceID int32) (r *AllianceInfoModel, err error) {
@@ -81,7 +81,7 @@ func (r *AllianceInfoModel) DataToAllianceInfoPb() *MazeFamily.AllianceInfo {
 		AllianceName:       proto.String(r.AllianceName),
 		AllianceCountLimit: proto.Int32(r.AllianceCountLimit),
 		FamilyIds:          r.FamilyIDs,
-		AllianceGroupId:    proto.Int32(r.AllianceGroupID),
+		AllianceGroupId:    proto.Int64(r.AllianceGroupID),
 	}
 }
 
@@ -100,7 +100,7 @@ func (r *AllianceInfoModel) RemoveFamilyID(ctx context.Context, familyID int32) 
 	return nil
 }
 
-func (r *AllianceInfoModel) SetAllianceGroupID(ctx context.Context, allianceGroupID int32) error {
+func (r *AllianceInfoModel) SetAllianceGroupID(ctx context.Context, allianceGroupID int64) error {
 	if r.AllianceGroupID != 0 {
 		return nil
 	}
