+++ model/familymodel/family_info_model.go
@@ -63,7 +63,7 @@ type FamilyInfoModel struct {
 	MemberCountLimit int32           `json:"member_count_limit,omitempty"`
 	FamilyMembers    []*FamilyMember `json:"family_members,omitempty"`
 	FamilyApplyUsers []*FamilyMember `json:"family_apply_users,omitempty"`
-	FamilyGroupID    int32           `json:"group_id,omitempty"`
+	FamilyGroupID    int64           `json:"group_id,omitempty"`
 }
 
 type FamilysInfoModel []*FamilyInfoModel
@@ -252,7 +252,7 @@ func (r *FamilyInfoModel) DataToFamilyInfoPb(ctx context.Context) *MazeFamily.Fa
 		FamilySetting: &MazeFamily.FamilySetting{
 			FamilyJoinType: MazeFamily.FamilyJoinType(r.FamilySetting.FamilyJoinType).Enum(),
 		},
-		GroupId: proto.Int32(r.FamilyGroupID),
+		GroupId: proto.Int64(r.FamilyGroupID),
 	}
 }
 
@@ -406,7 +406,7 @@ func (r *FamilyInfoModel) CheckUserInApplyList(ctx context.Context, userID uint6
 }
 
 // 设置家族群组ID
-func (r *FamilyInfoModel) SetFamilyGroupID(ctx context.Context, groupID int32) error {
+func (r *FamilyInfoModel) SetFamilyGroupID(ctx context.Context, groupID int64) error {
 	if r.FamilyGroupID != 0 {
 		return errors.New("family group id has set")
 	}
