+++ services/allianceservice/alliance_add.go
@@ -24,7 +24,7 @@ func (s *service) CreateAlliance(ctx context.Context, allianceName string) error
 	allianceID := allianceListModel.GetAllianceID(ctx)
 
 	allianceInfoModel := alliancemodel.NewAllianceInfoModel(ctx, allianceID, allianceName)
-	allianceInfoModel.SetAllianceGroupID(ctx, int32(time.Now().Unix()))
+	allianceInfoModel.SetAllianceGroupID(ctx, int64(time.Now().Unix()))
 	err = allianceInfoModel.Save(ctx)
 	if err != nil {
 		logger.CtxError(ctx, "AddAlliance Save allianceInfoModel err", zap.Error(err))
